Initial version of timing metrics in upload digest workflow
