+++ pstatus-notifications-workflow-ktor/src/main/kotlin/query/UploadDigestCountsQuery.kt
@@ -1,0 +1,87 @@
+package gov.cdc.ocio.processingnotifications.query
+
+import gov.cdc.ocio.database.QueryBuilder
+import gov.cdc.ocio.database.persistence.ProcessingStatusRepository
+import gov.cdc.ocio.processingnotifications.model.UploadDigestResponse
+import gov.cdc.ocio.types.InstantRange
+import java.time.LocalDate
+
+
+class UploadDigestCountsQuery(
+    repository: ProcessingStatusRepository
+): QueryBuilder(repository.reportsCollection) {
+
+    private fun build(
+        utcDateToRun: LocalDate,
+        dataStreamIds: List<String>,
+        dataStreamRoutes: List<String>,
+        jurisdictions: List<String>
+    ): String {
+        val instantRange = InstantRange.fromLocalDate(utcDateToRun)
+        val startTime = timeFunc(instantRange.start.epochSecond)
+        val endTime = timeFunc(instantRange.endInclusive.epochSecond)
+
+        val jurisdictionIdsList = listForQuery(jurisdictions)
+        val dataStreamIdsList = listForQuery(dataStreamIds)
+        val dataStreamRoutesList = listForQuery(dataStreamRoutes)
+
+        val querySB = StringBuilder()
+
+        querySB.append("""
+            SELECT ${cPrefix}dataStreamId, ${cPrefix}dataStreamRoute, ${cPrefix}jurisdiction
+            FROM $collectionName $cVar
+            WHERE ${cPrefix}stageInfo.action = 'upload-completed' AND ${cPrefix}stageInfo.status = 'SUCCESS'
+            """)
+
+        if (dataStreamIds.isNotEmpty())
+            querySB.append("""
+                AND ${cPrefix}dataStreamId IN ${openBkt}$dataStreamIdsList${closeBkt}
+                """)
+
+        if (dataStreamRoutesList.isNotEmpty())
+            querySB.append("""
+                AND ${cPrefix}dataStreamRoute IN ${openBkt}$dataStreamRoutesList${closeBkt}
+                """)
+
+        if (jurisdictionIdsList.isNotEmpty())
+            querySB.append("""
+                AND ${cPrefix}jurisdiction IN ${openBkt}$jurisdictionIdsList${closeBkt}
+                """)
+
+        querySB.append("""
+            AND ${cPrefix}dexIngestDateTime >= $startTime
+            AND ${cPrefix}dexIngestDateTime < $endTime
+            """)
+
+        return querySB.toString().trimIndent()
+    }
+
+    /**
+     * The function which gets the digest counts query and sends it to the corresponding db collection.
+     *
+     * @param utcDateToRun LocalDate
+     * @param dataStreamIds List<String>
+     * @param dataStreamRoutes List<String>
+     * @param jurisdictions List<String>
+     */
+    fun run(
+        utcDateToRun: LocalDate,
+        dataStreamIds: List<String>,
+        dataStreamRoutes: List<String>,
+        jurisdictions: List<String>,
+    ): List<UploadDigestResponse> {
+        return runCatching {
+            val query = build(
+                utcDateToRun,
+                dataStreamIds,
+                dataStreamRoutes,
+                jurisdictions
+            )
+            logger.info("Upload digest counts query:\n$query")
+            return@runCatching collection.queryItems(query, UploadDigestResponse::class.java)
+        }.getOrElse {
+            logger.error("Error occurred while getting a digest of upload counts: ${it.localizedMessage}")
+            throw it
+        }
+    }
+}
